squashed eject bug where the eject button worked as a toggle. Squashed bug where rear rollers were running sporadically. conflicting instructions were the reason for both of these issues.
added debouncing to teleop; ineffective.

Added one ball in the hot goal script.
Added 2 balls in the hot goal script; has the ball start out in front, letting it travel faster. Fires while moving to save time.

chezy vision only uses the state of the left square. If it is "lit" (along the side, hand removed) the robot fires as soon as possible, if it is not, the robot waits for the 5 minute mark.